Consolidate all the files that create instance of each controller to a single file.
