--- app/controllers/app.js
@@ -1,3 +1,0 @@
-const AppController = require("./definitions/AppController");
-
-module.exports = new AppController();
