--- app/controllers/appType.js
@@ -1,3 +1,0 @@
-const AppTypeController = require('./definitions/AppTypeController');
-
-module.exports = new AppTypeController();
