--- app/controllers/auth.js
@@ -1,3 +1,0 @@
-const AuthController = require("./definitions/AuthController");
-
-module.exports = new AuthController();
