--- app/controllers/feed.js
@@ -1,3 +1,0 @@
-const FeedController = require("./definitions/FeedController");
-
-module.exports = new FeedController();
