--- app/controllers/feedentries.js
@@ -1,7 +1,0 @@
-const FeedEntryController = require("./definitions/FeedEntryController");
-
-module.exports = {
-    commentCtrl: new FeedEntryController("comment"),
-    elevateCtrl: new FeedEntryController("elevate"),
-    deelevateCtrl: new FeedEntryController("deelevate"),
-}
