--- app/controllers/invite.js
@@ -1,3 +1,0 @@
-const InviteController = require("./definitions/InviteController");
-
-module.exports = new InviteController();
