--- app/controllers/register.js
@@ -1,3 +1,0 @@
-const RegisterController = require("./definitions/RegisterController");
-
-module.exports = new RegisterController();
