--- app/controllers/room.js
@@ -1,3 +1,0 @@
-const RoomController = require('./definitions/RoomController');
-
-module.exports = new RoomController();
