--- app/controllers/student.js
@@ -1,3 +1,0 @@
-const StudentController = require("./definitions/StudentController");
-
-module.exports = new StudentController();
