--- app/controllers/token.js
@@ -1,3 +1,0 @@
-const TokenController = require("./definitions/TokenController");
-
-module.exports = new TokenController();
