--- app/controllers/user.js
@@ -1,3 +1,0 @@
-const UserController = require("./definitions/UserController");
-
-module.exports = new UserController();
