--- app/controllers/validateEmail.js
@@ -1,3 +1,0 @@
-const ValidateEmailController = require("./definitions/ValidateEmailController");
-
-module.exports = new ValidateEmailController();
